+++ NEWS.rst
@@ -13,7 +13,8 @@ parallel_stereo (:numref:`parallel_stereo`):
     (:numref:`kh9`).
   * Added a status file that shows the progress in processing tiles
     (:numref:`stereo_diag`).
-  
+  * Added an output shapefile that shows the tiles being processed
+    (:numref:`stereo_diag`).
   * Left and right alignment matrices are now saved in plain text format. Older
     .exr files are still read. Support for them will be removed in the next
     release (:numref:`outputfiles`).